agente/requirements.txt
Flask
python-dotenv
google-generativeai
pandas
scikit-learn
